% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthIntegrationFunctions.R
\name{calcDWatAgeAtConstantTemp}
\alias{calcDWatAgeAtConstantTemp}
\title{Calculate time series of dry weight (DW)-at-age for a constant temperature.}
\usage{
calcDWatAgeAtConstantTemp(n, DW, T, durationFcn, growthFcn)
}
\arguments{
\item{n}{- number of integration time steps}

\item{DW}{- initial dry weight (units are arbitrary)}

\item{T}{- temperature (deg C)}

\item{durationFcn}{- temperature-dependent function giving duration}

\item{growthFcn}{- temperature-dependent function giving DW growth rate}
}
\value{
tibble with columns age-in-stage,temperature, and DW
}
\description{
Function to calculate time series of DW-at-age at a constant temperature.
}
\details{
Uses \code{\link{integrateGrowthDW}} to calculate a time series of DW-at-age given
temperature-dependent duration and DW growth functions.
}
\examples{
## initDW = 0.01; #--milligrams
## tbl1 = calcDWatAgeAtConstantTemp(100,initDW,5,egg_StageDuration,egg_GrowthRateDW);

##  initDW = 0.14; #--milligrams
## tbl2 = calcDWatAgeAtConstantTemp(100,initDW,5,ysl_YolkSacAbsorption,ysl_GrowthRateDW)

}
