% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthIntegrationFunctions.R
\name{calcSLatAgeAtConstantTemp}
\alias{calcSLatAgeAtConstantTemp}
\title{Calculate SL-at-age for a constant temperature.}
\usage{
calcSLatAgeAtConstantTemp(n, SL, T, durationFcn, growthFcn)
}
\arguments{
\item{n}{- number of integration time steps}

\item{SL}{- initial standard length (mm)}

\item{T}{- temperature (deg C)}

\item{durationFcn}{- temperature-dependent function giving duration}

\item{growthFcn}{- temperature-dependent function giving growth rate}
}
\value{
tibble with columns age and SL
}
\description{
Function to calculate SL-at-age at a constant temperature.
}
\details{
Uses \code{\link{integrateGrowthSL}} to calculate a time series of size-at-age given
temperature-dependent duration and growth functions.
}
