% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EggStage.R
\name{egg_CalcHatchSuccess}
\alias{egg_CalcHatchSuccess}
\title{Calculate the hatch success of Pacific cod eggs by temperature}
\usage{
egg_CalcHatchSuccess(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
value
}
\description{
Function to calculate the hatch success of Pacific cod eggs by temperature.
}
\details{
Based on Laurel and Rogers (2020)
}
