% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EggStage.R
\name{egg_GrowthRateDW}
\alias{egg_GrowthRateDW}
\title{Calculate the temperature-dependent growth rate (DW) for Pacific cod embryos}
\usage{
egg_GrowthRateDW(T)
}
\arguments{
\item{T}{- vector of temperatures (deg C)}
}
\value{
vector of corresponding growth rates (g/g/d)
}
\description{
Function to calculate the temperature-dependent growth rate (DW) for Pacific cod embryos.
}
\details{
Based on Hurst et al. 2010.
}
