% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EggStage.R
\name{egg_PlotHatchSuccess}
\alias{egg_PlotHatchSuccess}
\title{Plot the hatch success of Pacific cod eggs by temtperature}
\usage{
egg_PlotHatchSuccess(T = seq(from = 0, to = 15, by = 0.1))
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
ggplot2 objct
}
\description{
Function to plot the hatch success of Pacific cod eggs by temtperature.
}
\details{
Based on Laurel and Rogers (2020)
}
