% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EggStage.R
\name{egg_StageDuration}
\alias{egg_StageDuration}
\title{Calculate the temperature-dependent duration of the Pacific cod egg stage}
\usage{
egg_StageDuration(T)
}
\arguments{
\item{T}{- vector of temperatures (deg C)}
}
\value{
vector of corresponding stage durations (days)
}
\description{
Function to calculate the temperature-dependent duration of the Pacific cod egg stage.
}
\details{
Based on Hurst et al. 2010.
}
