% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs_HatchSuccess.R
\name{eggs_CalcHatchSuccess}
\alias{eggs_CalcHatchSuccess}
\title{Calculate the hatch success of Pacific cod eggs by temtperature}
\usage{
eggs_CalcHatchSuccess(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
value
}
\description{
Function to calculate the hatch success of Pacific cod eggs by temtperature.
}
\details{
Based on Laurel and Rogers (2020)
}
