% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_MapGrowthRate.R
\name{gg_MapGrowthRate}
\alias{gg_MapGrowthRate}
\title{Create a \pkg{ggplot2}-style map of growth rates}
\usage{
gg_MapGrowthRate(
  sf_dfr = NULL,
  rate_col = "growth_rate",
  label = "growth rate (1/day)",
  scale_fill = ggplot2::scale_fill_viridis_c(option = "A"),
  basemap = NULL,
  rateFcn = postflexionLarvae_GrowthRate_WW,
  roms_grid = wtsROMS::getGrid("CGOA"),
  roms_mo = "~/Work/Projects/ROMS/CGOA/cgoa_avg_0005.nc",
  timeslice = 1,
  vertical_layer = 1
)
}
\arguments{
\item{sf_dfr}{- \pkg{sf}-style dataframe with rate information (or NULL)}

\item{rate_col}{- name of column with rate information (default="growth_rate")}

\item{label}{- label for map legend (default = "growth rate (1/day)")}

\item{basemap}{- \pkg{ggplot2}-style basemap to plot growth rates on top of}

\item{rateFcn}{- temperature-dependent rate function to use (if \code{sf_dfr} is NULL)}

\item{roms_grid}{- sf dataframe representing a ROMS grid object (if \code{sf_dfr} is NULL)}

\item{roms_mo}{- sf dataframe representing a ROMS model object, or path to a ROMS model netcdf file  (if \code{sf_dfr} is NULL)}

\item{timeslice}{- time slice to extract (if roms_mo is a file path) (if \code{sf_dfr} is NULL)}

\item{vertical_layer}{- vertical layer to extract (if roms_mo is a file path) (if \code{sf_dfr} is NULL)}

\item{palette}{- palette for growth info (default = NULL to use palette of basemap)}
}
\value{
\pkg{ggplot2}-style map of growth rates
}
\description{
Function to create a \pkg{ggplot2}-style map of growth rates.
}
\details{
If sf_dfr is NULL, then spatially-explicit growth rates are calculated using the \code{rateFcn}
and associated information, then mapped.

 @note This function can also map other spatially-explicit values by correctly specifying the
 \code{rateCol} for the input \code{sf_dfr}.
}
