% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthRates_Compare.R
\name{growthRates_CompareAllStages}
\alias{growthRates_CompareAllStages}
\title{Plot growth rates as functions of temperature for all life stages}
\usage{
growthRates_CompareAllStages(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
list of ggplot2 plots for growth rates vs. temperature
}
\description{
Function to plot growth rates as functions of temperature for all life stages.
}
\details{
From Hurst et al., 2010.
#'
}
