% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthRates_Compare.R
\name{growthRates_CompareJuveniles}
\alias{growthRates_CompareJuveniles}
\title{Compare growth rates as functions of temperature for benthic juveniles}
\usage{
growthRates_CompareJuveniles(T, W = 6.8, EDc = 4138)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
ggplot2 plot for growth rates vs. temperature
}
\description{
Function to compare growth rates as functions of temperature for benthic juveniles.
}
\details{
From Hurst et al., 2010 and Hurst et al. 2018.
#'
}
