% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juveniles_BioEnAllometries.R
\name{juv_Ca}
\alias{juv_Ca}
\title{Calculate the allometric component of consumption}
\usage{
juv_Ca(w, P = 1, AC = 0.063, BC = -0.07, EDp = 4113)
}
\arguments{
\item{w}{- weight (g)}

\item{P}{- fraction of maximum consumption actually consumed}

\item{AC}{- scale in allometric relationship}

\item{BC}{- weight exponent in allometric equation}

\item{EDp}{- energy density of prey [J/g prey]}
}
\value{
value (in J/g fish/d)
}
\description{
Function to calculate the allometric component of consumption.
}
\details{
Based on allometric part of eq. S3 and values in Table 1
in the Supplement to Hurst et al. 2018:\cr
\eqn{  Cg = P * AC * w^BC }   [g prey/g fish/d]
\eqn{  Cj = Cg * EDp}         [   J/g fish/d]
}
\seealso{
Other juvenile bioenergetic functions: 
\code{\link{juvBioenergetics_CompareRelationships}()},
\code{\link{juvBioenergetics_PlotRelationships}()},
\code{\link{juv_BioEnGrowth}()},
\code{\link{juv_CT}()},
\code{\link{juv_GT}()},
\code{\link{juv_MT}()},
\code{\link{juv_Ra}()},
\code{\link{juv_fC}()},
\code{\link{juv_fR}()}
}
\concept{juvenile bioenergetic functions}
