% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juveniles_BioEnFunctionalResponses.R
\name{juv_fC}
\alias{juv_fC}
\title{Calculate the quadratic functional response to temperature in the consumption relationship}
\usage{
juv_fC(t, a = -0.0076, b = 0.1859, c = -0.1424)
}
\arguments{
\item{t}{- temperature (deg C)}

\item{a}{- quadratic coefficient in quadratic equation}

\item{b}{- linear coefficient in quadratic equation}

\item{c}{- constant coefficient in quadratic equation}
}
\value{
value
}
\description{
Function to calculate the quadratic functional response to temperature in the consumption relationship.
}
\details{
Based on eq. S10
in the Supplement to Hurst et al. 2018\cr
\eqn{val = a*t^2+b*t+c}  [unitless]
}
\seealso{
Other juvenile bioenergetic functions: 
\code{\link{juvBioenergetics_CompareRelationships}()},
\code{\link{juvBioenergetics_PlotRelationships}()},
\code{\link{juv_BioEnGrowth}()},
\code{\link{juv_CT}()},
\code{\link{juv_C}()},
\code{\link{juv_GT}()},
\code{\link{juv_MT}()},
\code{\link{juv_Ra}()},
\code{\link{juv_fR}()}
}
\concept{juvenile bioenergetic functions}
