% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juveniles_BioEnObservedRates.R
\name{juv_GT}
\alias{juv_GT}
\title{Calculate observed growth (J/g fish/day) of a (nominally) 6.8 g fish as function of temperature}
\usage{
juv_GT(T, EDc = 4138)
}
\arguments{
\item{T}{- temperature (deg C)}

\item{EDc}{- energy density of Pacific cod juveniles (J/g)}
}
\value{
value for growth rate (fraction wet weight/day) of juvenile Pacific cod weighing (nominally) 6.8 g.
}
\description{
Function to calculate observed growth rate (J/g fish/day) of a (nominally) 6.8 g fish as a function of temperature.
}
\details{
Based on eq. S8 (from Laurel et al. 2016, in % body mass/day) as
given in Supplement to Hurst et al. 2018: \cr
\itemize{
  \item{\eqn{ GT = [(0.2494 + 0.3216*T - 0.0069*T^2 - 0.0004*T^3)/100]*EDc},[J/g fish/day]}
}
}
\seealso{
Other juvenile bioenergetic functions: 
\code{\link{juvBioenergetics_CompareRelationships}()},
\code{\link{juvBioenergetics_PlotRelationships}()},
\code{\link{juv_BioEnGrowth}()},
\code{\link{juv_CT}()},
\code{\link{juv_C}()},
\code{\link{juv_MT}()},
\code{\link{juv_Ra}()},
\code{\link{juv_fC}()},
\code{\link{juv_fR}()}
}
\concept{juvenile bioenergetic functions}
