% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juveniles_BioEnObservedRates.R
\name{juv_MT}
\alias{juv_MT}
\title{Calculate functional reponse of respiration to temperature}
\usage{
juv_MT(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
value for functional response (units??)
}
\description{
Function to calculate functional reponse of respiration to temperature.
}
\details{
Based on eq. S5 (from Oh et al. 2010) as
given in Supplement to Hurst et al. 2018:
\itemize{
  \item{\eqn{mt = (-1.04e-5)*T^2 + (3.38e-4)*T + (-1.02e-3)}, units??}
}
}
\seealso{
Other juvenile bioenergetic functions: 
\code{\link{juvBioenergetics_CompareRelationships}()},
\code{\link{juvBioenergetics_PlotRelationships}()},
\code{\link{juv_BioEnGrowth}()},
\code{\link{juv_CT}()},
\code{\link{juv_C}()},
\code{\link{juv_GT}()},
\code{\link{juv_Ra}()},
\code{\link{juv_fC}()},
\code{\link{juv_fR}()}
}
\concept{juvenile bioenergetic functions}
