% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juveniles_BioEnAllometries.R
\name{juv_Ra}
\alias{juv_Ra}
\title{Calculate the allometric component of the respiration relationship}
\usage{
juv_Ra(w, AM = 1, AR = 0.003, BR = -0.291, O = 13560)
}
\arguments{
\item{w}{- weight (g)}

\item{AM}{- adjustment for average metabolic activity}

\item{AR}{- scale in allometric relationship}

\item{BR}{- weight exponent in allometric equation}

\item{O}{- conversion for calorific equivalent of oxygen consumption (13560 J/g O2)}
}
\value{
value in J/g fish/d
}
\description{
Function to calculate the allometric component of the respiration relationship.
}
\details{
Based on allometric part of eq. S2 and values in Table 1
in the Supplement to Hurst et al. 2018: \cr
\eqn{Ro = AM * AR * w^BR}  [g O2/g fish/d]
\eqn{R = Ro * O}           [   J/g fish/d]
}
\seealso{
Other juvenile bioenergetic functions: 
\code{\link{juvBioenergetics_CompareRelationships}()},
\code{\link{juvBioenergetics_PlotRelationships}()},
\code{\link{juv_BioEnGrowth}()},
\code{\link{juv_CT}()},
\code{\link{juv_C}()},
\code{\link{juv_GT}()},
\code{\link{juv_MT}()},
\code{\link{juv_fC}()},
\code{\link{juv_fR}()}
}
\concept{juvenile bioenergetic functions}
