% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juveniles_BioEnCompareRelationships.R
\name{juvBioenergetics_PlotRelationships}
\alias{juvBioenergetics_PlotRelationships}
\title{Plot the bioenergetic relationships for juvenile Pacific cod}
\usage{
juvBioenergetics_PlotRelationships(
  W = seq(from = 3, to = 10, by = 0.2),
  T = seq(from = 0, to = 20, by = 2)
)
}
\arguments{
\item{W}{- weights, in g}

\item{T}{- temperatures, in deg C}
}
\value{
list of \code{ggplot2}-type plots
}
\description{
Function to plot the bioenergetic relationships for juvenile Pacific cod.
}
\seealso{
Other juvenile bioenergetic functions: 
\code{\link{juvBioenergetics_CompareRelationships}()},
\code{\link{juv_BioEnGrowth}()},
\code{\link{juv_CT}()},
\code{\link{juv_C}()},
\code{\link{juv_GT}()},
\code{\link{juv_MT}()},
\code{\link{juv_Ra}()},
\code{\link{juv_fC}()},
\code{\link{juv_fR}()}
}
\concept{juvenile bioenergetic functions}
