% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juveniles_Growth.R
\name{juveniles_GrowthRate_TL}
\alias{juveniles_GrowthRate_TL}
\title{Calculate growth rate (mm/day TL) for juvenile larvae}
\usage{
juveniles_GrowthRate_TL(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of growth rates
}
\description{
Function to calculate growth rate (mm/day SL) for juvenile larvae.
}
\details{
From Hurst et al., 2010. Valid range is 3-11 deg C.
}
