% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Juveniles.R
\name{juveniles_GrowthRateWW}
\alias{juveniles_GrowthRateWW}
\title{Calculate growth rate (1/day wet weight) for juvenile larvae}
\usage{
juveniles_GrowthRateWW(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of growth rates
}
\description{
Function to calculate growth rate (1/day wet weight) for juvenile larvae.
}
\details{
From Hurst et al., 2010. Valid range is 2-12 deg C.
}
