% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PostflexionLarvae.R
\name{postflexionLarvae_GrowthRateDW}
\alias{postflexionLarvae_GrowthRateDW}
\title{Calculate growth rate (1/day dry weight) for postflexion larvae}
\usage{
postflexionLarvae_GrowthRateDW(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of growth rates
}
\description{
Function to calculate growth rate (1/day dry weight) for postflexion larvae.
}
\details{
From Hurst et al., 2010. Valid range is 3-11 deg C.
}
