% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postflexionLarvae_Growth.R
\name{postflexionLarvae_MapGrowthRate_WW}
\alias{postflexionLarvae_MapGrowthRate_WW}
\title{Map growth rates (1/day wet weight) for postflexion larvae}
\usage{
postflexionLarvae_MapGrowthRate_WW(
  roms_grid = wtsROMS::getGrid("CGOA"),
  sf_mo = "~/Work/Projects/ROMS/CGOA/cgoa_avg_0005.nc",
  timeslice = 1,
  vertical_layer = 1,
  basemap = NULL
)
}
\arguments{
\item{roms_grid}{- sf dataframe representing a ROMS grid object}

\item{sf_mo}{- sf dataframe representing a ROMS model object, or path to a ROMS model netcdf file}

\item{timeslice}{- time slice to extract}

\item{vertical_layer}{- vertical layer to extract}

\item{basemap}{- basemap to plot growth rates on top of}
}
\value{
map of growth rates
}
\description{
Function to map growth rates (1/day wet weight) for postflexion larvae.
}
\details{
From Hurst et al., 2010. Valid range is 3-11 deg C.
}
