% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preflexionLarvae_Growth.R
\name{preflexionLarvae_GrowthRate_SL}
\alias{preflexionLarvae_GrowthRate_SL}
\title{Calculate growth rate (mm/day SL) for preflexion larvae}
\usage{
preflexionLarvae_GrowthRate_SL(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of growth rates
}
\description{
Function to calculate growth rate (mm/day SL) for preflexion larvae.
}
\details{
From Hurst et al., 2010. Valid range is 2-11 deg C.
}
