% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PreflexionLarvae.R
\name{preflexionLarvae_GrowthRateDW}
\alias{preflexionLarvae_GrowthRateDW}
\title{Calculate growth rate (1/day dry weight) for preflexion larvae}
\usage{
preflexionLarvae_GrowthRateDW(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of growth rates
}
\description{
Function to calculate growth rate (1/day dry weight) for preflexion larvae.
}
\details{
From Hurst et al., 2010. Valid range is 2-11 deg C.
}
