% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PreflexionLarvae.R
\name{preflexionLarvae_GrowthRateSL}
\alias{preflexionLarvae_GrowthRateSL}
\title{Calculate growth rate (mm/day SL) for preflexion larvae}
\usage{
preflexionLarvae_GrowthRateSL(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of growth rates
}
\description{
Function to calculate growth rate (mm/day SL) for preflexion larvae.
}
\details{
Corrected from Hurst et al., 2010. Valid range is 2-11 deg C.
}
