% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_CalcTdepGrowthRatesFromROMS.R
\name{sf_CalcTdepGrowthRatesFromROMS}
\alias{sf_CalcTdepGrowthRatesFromROMS}
\title{Get a pkg{sf} dataframe with spatially-explicit temperature-dependent growth rates from ROMS model output}
\usage{
sf_CalcTdepGrowthRatesFromROMS(
  rateFcn = postflexionLarvae_GrowthRate_WW,
  roms_grid = wtsROMS::getGrid("CGOA"),
  roms_mo = "~/Work/Projects/ROMS/CGOA/cgoa_avg_0005.nc",
  timeslice = 1,
  vertical_layer = 1,
  crs = NULL
)
}
\arguments{
\item{rateFcn}{- temperature-dependent rate function to use}

\item{roms_grid}{- sf dataframe representing a ROMS grid object}

\item{roms_mo}{- sf dataframe representing a ROMS model object, or path to a ROMS model netcdf file}

\item{timeslice}{- time slice to extract (if roms_mo is a file path)}

\item{vertical_layer}{- vertical layer to extract (if roms_mo is a file path)}

\item{crs}{- \pkg{sf}-style coordinate reference system object, or NULL (default)}
}
\value{
\pkg{sf} dataframe with spataly-explicit temperature-dependent growth rates
}
\description{
Function to get a pkg{sf} dataframe with spatially-explicit temperature-dependent growth rates from ROMS model output.
}
\details{
Output dataframe will be in the coordinate reference system of the ROMS model
grid unless the input \code{crs} was not NULL (the default).

 @note This function can also calculate spatially-explicit egg hatch success if \code{egg_HatchSuccess}
 is given as the \code{rateFcn}. The "growth_rate" column of the resulting \pkg{sf} dataframe
 will contain the hatch success information.
}
