% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdgGrowthRates.R
\name{stdgGrowthRateDW_NonEggStages}
\alias{stdgGrowthRateDW_NonEggStages}
\title{Calculate STDG growth rates (1/day wet weight) for post-egg stages}
\usage{
stdgGrowthRateDW_NonEggStages(T, M)
}
\arguments{
\item{T}{- vector of temperature (deg C)}

\item{M}{- vector of dry weight (micrograms))}
}
\value{
vector of growth rates
}
\description{
Function to calculate STDG growth rates (1/day wet weight) for post-egg stages.
}
\details{
From Hurst et al., 2010. Valid range is 3-11 deg C.
T and M should have the same length, unless one has length 1.
}
