% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdgGrowthRates.R
\name{stdgGrowthRateSL_EggStages}
\alias{stdgGrowthRateSL_EggStages}
\title{Calculate STDG growth rates (mm/day SL) for egg stages}
\usage{
stdgGrowthRateSL_EggStages(T, L)
}
\arguments{
\item{T}{- vector of temperature (deg C)}

\item{L}{- vector of standard length (mm)}
}
\value{
vector of growth rates
}
\description{
Function to calculate STDG growth rates (mm/day SL) for egg stages.
}
\details{
From Hurst et al., 2010. Valid range is 3-11 deg C.
T and L should have the same length, unless one has length 1.

Note that these growth rates are independent of L.
}
