% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdgGrowthRates.R
\name{stdgGrowthRateSL_NonEggStages}
\alias{stdgGrowthRateSL_NonEggStages}
\title{Calculate STDG growth rates (mm/day SL) for non-egg stages}
\usage{
stdgGrowthRateSL_NonEggStages(T, L)
}
\arguments{
\item{T}{- vector of temperatures (deg C)}

\item{L}{- vector of standard lengths (mm)}
}
\value{
vector of growth rates
}
\description{
Function to calculate STDG growth rates (mm/day SL) for non-egg stages.
}
\details{
From Hurst et al., 2010. Valid range is 3-11 deg C.
T and M should have the same length, unless one has length 1.
}
