% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_YSL.R
\name{ysl_ConvertSLtoDW}
\alias{ysl_ConvertSLtoDW}
\title{Calculate dry weight (mg)) corresponding to standard length (mm) for yolk-sac larvae}
\usage{
ysl_ConvertSLtoDW(sl)
}
\arguments{
\item{SL}{- vector of standard lengths (mm)}
}
\value{
vector of dry weights based on regression equation
}
\description{
Function to calculate dry weight (mg)) corresponding to standard length (mm) for yolk-sac larvae.
}
\details{
Based on data from Hurst et al., 2010
}
