% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysl_Growth.R
\name{ysl_GrowthRate_SL}
\alias{ysl_GrowthRate_SL}
\title{Calculate growth rate (mm/day SL) for yolk-sac larvae}
\usage{
ysl_GrowthRate_SL(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of growth rates
}
\description{
Function to calculate growth rate (mm/day SL) for yolk-sac larvae.
}
\details{
From Hurst et al., 2010. Valid range is 2-11 deg C.
}
