% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_YSL.R
\name{ysl_GrowthRateDW}
\alias{ysl_GrowthRateDW}
\title{Calculate growth rate (1/day dry weight) for yolk-sac larvae}
\usage{
ysl_GrowthRateDW(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of growth rates
}
\description{
Function to calculate growth rate (1/day dry weight) for yolk-sac larvae.
}
\details{
From Hurst et al., 2010. Valid range is 2-11 deg C.
}
