% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_YSL.R
\name{ysl_GrowthRateSL}
\alias{ysl_GrowthRateSL}
\title{Calculate growth rate (mm/day SL) for yolk-sac larvae}
\usage{
ysl_GrowthRateSL(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of growth rates
}
\description{
Function to calculate growth rate (mm/day SL) for yolk-sac larvae.
}
\details{
Corrected version from Hurst et al., 2010. Valid range is 2-11 deg C.
}
