% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysl_Growth.R
\name{ysl_PNR}
\alias{ysl_PNR}
\title{Calculate the "point-of-no-return" for non-feeding yolk-sac larvae after yolk-sac absorption}
\usage{
ysl_PNR(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of times to point-of-no-return for non-feeding YSL.
}
\description{
Function to calculate the "point-of-no-return" for
non-feeding yolk-sac larvae after yolk-sac absorption.
}
\details{
None.
}
