% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_YSL.R
\name{ysl_SwimmingSpeed}
\alias{ysl_SwimmingSpeed}
\title{Calculate the swimming speed of YSL-stage larvae}
\usage{
ysl_SwimmingSpeed(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of times to point-of-no-return for non-feeding YSL.
}
\description{
Function to calculate the the swimming speed of YSL-stage larvae.
}
\details{
None.
}
