% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_YSL.R
\name{ysl_YolkSacAbsorption}
\alias{ysl_YolkSacAbsorption}
\title{Calculate time to yolk-sac absorption for yolk-sac larvae}
\usage{
ysl_YolkSacAbsorption(T)
}
\arguments{
\item{T}{- temperature (deg C)}
}
\value{
vector of times to yolk sac absorption
}
\description{
Function to calculate time to yolk-sac absorption for yolk-sac larvae.
}
\details{
None.
}
