% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLifeStageInfo.PCod.R
\name{getLifeStageInfo.PCod}
\alias{getLifeStageInfo.PCod}
\title{Get a list that defines life stage info for the DisMELS Pacific cod IBM.}
\usage{
getLifeStageInfo.PCod(resType = "NEW2.0SC")
}
\arguments{
\item{resType}{- results file type (i.e., 'NEW2.0SC')}
}
\value{
a list
}
\description{
Function to get a list that defines life stage info for the DisMELS Pacific cod IBM.
}
\details{
Uses \code{rDisMELS::getStandardAttributes}.
}
